+++ src/pages/concept-registration-page/form-concept/form-allowed-and-discouraged-term/form-allowed-and-discouraged-term-pure.component.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Field } from 'formik';
-import { InputTagsField } from '../../../components/field-input-tags/field-input-tags.component';
-import { HelpText } from '../../../components/help-text/help-text.component';
-import { localization } from '../../../lib/localization';
+import { InputTagsField } from '../../../../components/field-input-tags/field-input-tags.component';
+import { HelpText } from '../../../../components/help-text/help-text.component';
+import { localization } from '../../../../lib/localization';
 
 export const FormAllowedAndDiscouragedPure = (): JSX.Element => (
   <div>
