+++ src/pages/concept-registration-page/form-concept/form-use-of-concept/form-useOfConcept-pure.component.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Field } from 'formik';
-import { InputField } from '../../../components/field-input/field-input.component';
-import { InputTagsField } from '../../../components/field-input-tags/field-input-tags.component';
-import { TextAreaField } from '../../../components/field-textarea/field-textarea.component';
-import { HelpText } from '../../../components/help-text/help-text.component';
-import { localization } from '../../../lib/localization';
+import { InputField } from '../../../../components/field-input/field-input.component';
+import { InputTagsField } from '../../../../components/field-input-tags/field-input-tags.component';
+import { TextAreaField } from '../../../../components/field-textarea/field-textarea.component';
+import { HelpText } from '../../../../components/help-text/help-text.component';
+import { localization } from '../../../../lib/localization';
 
 export const FormUseOfTermPure = (): JSX.Element => (
   <div>
